pulled over ty's refactor of mqtt stuff from https://github.com/vhs/vhs-laser-access/blob/feat/2024-improvements/lib/mqtt.js (just found this)
