+++ src/laserAccess.js
@@ -18,12 +18,11 @@ const EventEmitter = require('events').EventEmitter
 const Bluebird = require('bluebird')
 const CryptoJS = require('crypto-js')
 const debug = require('debug')('laser:control')
-const mqtt = require('mqtt')
 const rp = require('request-promise')
-
 const config = require('../config')
-const { writeSync } = require('fs')
 const Led = require('./led').Led
+
+const { maintenanceStatus } = require('./mqtt')
 
 const gpios = {
   GPIO_LASER: 22,
@@ -48,38 +47,6 @@ const LEDs = {
 }
 
 const emitter = new EventEmitter()
-
-// MQTT setup
-const mqttClient = mqtt.connect(
-  config.mqttServer || 'mqtt://127.0.0.1',
-  config.mqttOptions
-) // VHS Public Facing MQTT. Should eventually be changed to a private mqtt instance.
-const mqttTopic = config.mqttTopic || 'laser/maintenance'
-let maintenanceStatus = 'ok' // Default status is 'ok'
-
-mqttClient.on('connect', () => {
-  debug('Connected to MQTT broker')
-
-  mqttClient.subscribe(mqttTopic, (err) => {
-    if (err) {
-      console.error(`Failed to subscribe to topic: ${mqttTopic}`)
-    } else {
-      debug(`Subscribed to MQTT topic: ${mqttTopic}`)
-    }
-  })
-})
-
-mqttClient.on('message', (topic, message) => {
-  if (topic === mqttTopic) {
-    maintenanceStatus = message.toString() // Read the message and store it
-
-    debug(`Received message on ${mqttTopic}: ${maintenanceStatus}`)
-  }
-})
-
-mqttClient.on('error', (err) => {
-  console.error('MQTT error:', err)
-})
 
 LEDs.red.enable()
